accessible/tests/mochitest/relations/test_tabbrowser.xhtml

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>

<window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        title="Accessible XUL tabbrowser relation tests">

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js" />
  <script src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>

  <script type="application/javascript"
          src="../common.js" />
  <script type="application/javascript"
          src="../role.js" />
  <script type="application/javascript"
          src="../relations.js" />
  <script type="application/javascript"
          src="../events.js" />
  <script type="application/javascript"
          src="../browser.js"></script>

  <script type="application/javascript">
  <![CDATA[
    async function doTest() {
      // Load documents into tabs and wait for DocLoadComplete events caused by
      // these documents load before we start the test.

      let loaded = waitForEvents([
        [EVENT_DOCUMENT_LOAD_COMPLETE, e => e.DOMNode == tabDocumentAt(0)],
        [EVENT_DOCUMENT_LOAD_COMPLETE, e => e.DOMNode == tabDocumentAt(1)]
      ]);

      const docURIs = ["about:license", "about:mozilla"];
      tabBrowser().loadTabs(docURIs, {
        inBackground: false,
        replace: true,
        triggeringPrincipal:
          Services.scriptSecurityManager.getSystemPrincipal(),
      });
      // Flush layout, so as to guarantee that the a11y tree is constructed.
      browserDocument().documentElement.getBoundingClientRect();

      await loaded;

      ////////////////////////////////////////////////////////////////////////
      // 'labelled by'/'label for' relations for xul:tab and xul:tabpanel

      const tabs = Array.from(tabBrowser().tabContainer.allTabs);
      // For preloaded tabs, there might be items in this array where this relation
      // doesn't hold, so just deal with that:
      const panels = tabs.map(t => t.linkedBrowser.closest("tabpanels > *"));

      testRelation(panels[0], RELATION_LABELLED_BY, tabs[0]);
      testRelation(tabs[0], RELATION_LABEL_FOR, panels[0]);
      testRelation(panels[1], RELATION_LABELLED_BY, tabs[1]);
      testRelation(tabs[1], RELATION_LABEL_FOR, panels[1]);

      closeBrowserWindow();
      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    openBrowserWindow(doTest);
  ]]>
  </script>

  <vbox flex="1" style="overflow: auto;">
    <body xmlns="http://www.w3.org/1999/xhtml">
      <a target="_blank"
         href="https://bugzilla.mozilla.org/show_bug.cgi?id=552944"
         title="No relationship between tabs and associated property page in new tabbrowser construct">
        Mozilla Bug 552944
      </a><br/>
      <p id="display"></p>
      <div id="content" style="display: none">
      </div>
      <pre id="test">
      </pre>
    </body>

    <vbox id="eventdump"></vbox>
  </vbox>

</window>
